# openapi/openapi.json
{"openapi":"3.0.3","info":{"title":"i-Store API","version":"1.0.0","description":"OpenAPI specification generated from Zod schemas."},"servers":[{"url":"/api"}],"security":[{"bearerAuth":[]}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"CustomReportQuestionShowIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"CustomReportChoiceOption":{"type":"object","properties":{"_id":{"type":"string"},"label":{"type":"string"},"value":{"type":"string"},"description":{"type":"string"},"index":{"type":"integer"}},"required":["_id","label","value"]},"CustomReportQuestion":{"oneOf":[{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["choice"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"options":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"label":{"type":"string"},"value":{"type":"string"},"description":{"type":"string"},"index":{"type":"integer"}},"required":["_id","label","value"]}},"allowOther":{"type":"boolean"},"otherAnswerType":{"type":"string","enum":["textField","numberField"]}},"required":["questionId","index","query","kind","options"]},{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["choiceMultiSelect"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"options":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"label":{"type":"string"},"value":{"type":"string"},"description":{"type":"string"},"index":{"type":"integer"}},"required":["_id","label","value"]}},"allowOther":{"type":"boolean"},"otherAnswerType":{"type":"string","enum":["textField","numberField"]},"minSelect":{"type":"integer"},"maxSelect":{"type":"integer"}},"required":["questionId","index","query","kind","options"]},{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["textField"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"placeholder":{"type":"string"},"minLength":{"type":"integer"},"maxLength":{"type":"integer"}},"required":["questionId","index","query","kind"]},{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["numberField"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"min":{"type":"number"},"max":{"type":"number"}},"required":["questionId","index","query","kind"]}]},"CustomReportLatestReport":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string","nullable":true},"prvilege":{"type":"string"},"SecondPrivileage":{"type":"string","nullable":true},"interval":{"type":"object","nullable":true,"properties":{"type":{"type":"string","enum":["daily","weekly","monthly","yearly"]},"times":{"type":"array","items":{"type":"integer"}}},"required":["type"]},"status":{"type":"string","enum":["draft","published","archived"]},"version":{"type":"integer"},"questions":{"type":"array","items":{"oneOf":[{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["choice"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"options":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"label":{"type":"string"},"value":{"type":"string"},"description":{"type":"string"},"index":{"type":"integer"}},"required":["_id","label","value"]}},"allowOther":{"type":"boolean"},"otherAnswerType":{"type":"string","enum":["textField","numberField"]}},"required":["questionId","index","query","kind","options"]},{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["choiceMultiSelect"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"options":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"label":{"type":"string"},"value":{"type":"string"},"description":{"type":"string"},"index":{"type":"integer"}},"required":["_id","label","value"]}},"allowOther":{"type":"boolean"},"otherAnswerType":{"type":"string","enum":["textField","numberField"]},"minSelect":{"type":"integer"},"maxSelect":{"type":"integer"}},"required":["questionId","index","query","kind","options"]},{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["textField"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"placeholder":{"type":"string"},"minLength":{"type":"integer"},"maxLength":{"type":"integer"}},"required":["questionId","index","query","kind"]},{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["numberField"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"min":{"type":"number"},"max":{"type":"number"}},"required":["questionId","index","query","kind"]}]}},"publishedAt":{"type":"integer"}},"required":["id","title","prvilege","interval","status","version","questions"]},"CustomReportInterval":{"type":"object","properties":{"type":{"type":"string","enum":["daily","weekly","monthly","yearly"]},"times":{"type":"array","items":{"type":"integer"}}},"required":["type"]},"CustomReportListResponse":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string","nullable":true},"prvilege":{"type":"string"},"SecondPrivileage":{"type":"string","nullable":true},"interval":{"type":"object","nullable":true,"properties":{"type":{"type":"string","enum":["daily","weekly","monthly","yearly"]},"times":{"type":"array","items":{"type":"integer"}}},"required":["type"]},"status":{"type":"string","enum":["draft","published","archived"]},"version":{"type":"integer"},"questions":{"type":"array","items":{"oneOf":[{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["choice"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"options":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"label":{"type":"string"},"value":{"type":"string"},"description":{"type":"string"},"index":{"type":"integer"}},"required":["_id","label","value"]}},"allowOther":{"type":"boolean"},"otherAnswerType":{"type":"string","enum":["textField","numberField"]}},"required":["questionId","index","query","kind","options"]},{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["choiceMultiSelect"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"options":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"label":{"type":"string"},"value":{"type":"string"},"description":{"type":"string"},"index":{"type":"integer"}},"required":["_id","label","value"]}},"allowOther":{"type":"boolean"},"otherAnswerType":{"type":"string","enum":["textField","numberField"]},"minSelect":{"type":"integer"},"maxSelect":{"type":"integer"}},"required":["questionId","index","query","kind","options"]},{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["textField"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"placeholder":{"type":"string"},"minLength":{"type":"integer"},"maxLength":{"type":"integer"}},"required":["questionId","index","query","kind"]},{"type":"object","properties":{"questionId":{"type":"string"},"index":{"type":"integer"},"query":{"type":"string"},"kind":{"type":"string","enum":["numberField"]},"helpText":{"type":"string"},"required":{"type":"boolean"},"showIf":{"type":"object","properties":{"questionId":{"type":"string"},"optionIdEquals":{"type":"string"},"exists":{"type":"boolean"}},"required":["questionId"]},"min":{"type":"number"},"max":{"type":"number"}},"required":["questionId","index","query","kind"]}]}},"publishedAt":{"type":"integer"}},"required":["id","title","prvilege","interval","status","version","questions"]}},"CustomReportCreateRequest":{"type":"object","properties":{"title":{"type":"string","minLength":1},"description":{"type":"string"},"prvilege":{"type":"string","enum":["admin","manager","staff"]},"SecondPrivileage":{"type":"string","enum":["super","call-center","regular","accountant","technician","hr","inventory","marketing-executive"]},"interval":{"type":"object","properties":{"type":{"type":"string","enum":["daily","weekly","monthly","yearly"]},"times":{"type":"array","items":{"type":"number"}}},"required":["type"]},"questions":{"type":"array","items":{"oneOf":[{"type":"object","properties":{"index":{"type":"integer","minimum":0},"query":{"type":"string","minLength":1},"kind":{"type":"string","enum":["choice"]},"helpText":{"type":"string"},"required":{"type":"boolean","default":false},"showIf":{"type":"object","properties":{"questionIndex":{"type":"integer","minimum":0},"choiceIndexEquals":{"type":"integer","minimum":0},"exists":{"type":"boolean"}},"required":["questionIndex"]},"options":{"type":"array","items":{"type":"object","properties":{"label":{"type":"string","minLength":1},"value":{"type":"string","minLength":1},"description":{"type":"string"}},"required":["label","value"]},"minItems":1},"allowOther":{"type":"boolean"},"otherAnswerType":{"type":"string","enum":["textField","numberField"]}},"required":["index","query","kind","options"]},{"type":"object","properties":{"index":{"type":"integer","minimum":0},"query":{"type":"string","minLength":1},"kind":{"type":"string","enum":["choiceMultiSelect"]},"helpText":{"type":"string"},"required":{"type":"boolean","default":false},"showIf":{"type":"object","properties":{"questionIndex":{"type":"integer","minimum":0},"choiceIndexEquals":{"type":"integer","minimum":0},"exists":{"type":"boolean"}},"required":["questionIndex"]},"options":{"type":"array","items":{"type":"object","properties":{"label":{"type":"string","minLength":1},"value":{"type":"string","minLength":1},"description":{"type":"string"}},"required":["label","value"]},"minItems":1},"allowOther":{"type":"boolean"},"otherAnswerType":{"type":"string","enum":["textField","numberField"]},"minSelect":{"type":"integer","minimum":0},"maxSelect":{"type":"integer","minimum":1}},"required":["index","query","kind","options"]},{"type":"object","properties":{"index":{"type":"integer","minimum":0},"query":{"type":"string","minLength":1},"kind":{"type":"string","enum":["textField"]},"helpText":{"type":"string"},"required":{"type":"boolean","default":false},"showIf":{"type":"object","properties":{"questionIndex":{"type":"integer","minimum":0},"choiceIndexEquals":{"type":"integer","minimum":0},"exists":{"type":"boolean"}},"required":["questionIndex"]},"placeholder":{"type":"string"},"minLength":{"type":"integer","minimum":0},"maxLength":{"type":"integer","minimum":0}},"required":["index","query","kind"]},{"type":"object","properties":{"index":{"type":"integer","minimum":0},"query":{"type":"string","minLength":1},"kind":{"type":"string","enum":["numberField"]},"helpText":{"type":"string"},"required":{"type":"boolean","default":false},"showIf":{"type":"object","properties":{"questionIndex":{"type":"integer","minimum":0},"choiceIndexEquals":{"type":"integer","minimum":0},"exists":{"type":"boolean"}},"required":["questionIndex"]},"min":{"type":"number"},"max":{"type":"number"}},"required":["index","query","kind"]}]}}},"required":["title","prvilege","SecondPrivileage","questions"]},"CustomReportCreateResponse":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"CustomReportPublishRequest":{"type":"object","properties":{"questions":{"type":"array","items":{"oneOf":[{"type":"object","properties":{"index":{"type":"integer","minimum":0},"query":{"type":"string","minLength":1},"kind":{"type":"string","enum":["choice"]},"helpText":{"type":"string"},"required":{"type":"boolean","default":false},"showIf":{"type":"object","properties":{"questionIndex":{"type":"integer","minimum":0},"choiceIndexEquals":{"type":"integer","minimum":0},"exists":{"type":"boolean"}},"required":["questionIndex"]},"options":{"type":"array","items":{"type":"object","properties":{"label":{"type":"string","minLength":1},"value":{"type":"string","minLength":1},"description":{"type":"string"}},"required":["label","value"]},"minItems":1},"allowOther":{"type":"boolean"},"otherAnswerType":{"type":"string","enum":["textField","numberField"]}},"required":["index","query","kind","options"]},{"type":"object","properties":{"index":{"type":"integer","minimum":0},"query":{"type":"string","minLength":1},"kind":{"type":"string","enum":["choiceMultiSelect"]},"helpText":{"type":"string"},"required":{"type":"boolean","default":false},"showIf":{"type":"object","properties":{"questionIndex":{"type":"integer","minimum":0},"choiceIndexEquals":{"type":"integer","minimum":0},"exists":{"type":"boolean"}},"required":["questionIndex"]},"options":{"type":"array","items":{"type":"object","properties":{"label":{"type":"string","minLength":1},"value":{"type":"string","minLength":1},"description":{"type":"string"}},"required":["label","value"]},"minItems":1},"allowOther":{"type":"boolean"},"otherAnswerType":{"type":"string","enum":["textField","numberField"]},"minSelect":{"type":"integer","minimum":0},"maxSelect":{"type":"integer","minimum":1}},"required":["index","query","kind","options"]},{"type":"object","properties":{"index":{"type":"integer","minimum":0},"query":{"type":"string","minLength":1},"kind":{"type":"string","enum":["textField"]},"helpText":{"type":"string"},"required":{"type":"boolean","default":false},"showIf":{"type":"object","properties":{"questionIndex":{"type":"integer","minimum":0},"choiceIndexEquals":{"type":"integer","minimum":0},"exists":{"type":"boolean"}},"required":["questionIndex"]},"placeholder":{"type":"string"},"minLength":{"type":"integer","minimum":0},"maxLength":{"type":"integer","minimum":0}},"required":["index","query","kind"]},{"type":"object","properties":{"index":{"type":"integer","minimum":0},"query":{"type":"string","minLength":1},"kind":{"type":"string","enum":["numberField"]},"helpText":{"type":"string"},"required":{"type":"boolean","default":false},"showIf":{"type":"object","properties":{"questionIndex":{"type":"integer","minimum":0},"choiceIndexEquals":{"type":"integer","minimum":0},"exists":{"type":"boolean"}},"required":["questionIndex"]},"min":{"type":"number"},"max":{"type":"number"}},"required":["index","query","kind"]}]},"minItems":1}},"required":["questions"]},"CustomReportPublishResponse":{"type":"object","properties":{"id":{"type":"string"},"version":{"type":"integer"}},"required":["id","version"]},"CustomReportSubmitResponseRequest":{"type":"object","properties":{"version":{"type":"integer","minimum":1},"answers":{"type":"array","items":{"type":"object","properties":{"questionId":{"type":"string"},"textValue":{"type":"string"},"numberValue":{"type":"number"},"choiceValue":{"type":"object","properties":{"optionIds":{"type":"array","items":{"type":"string"},"minItems":1},"otherText":{"type":"string"},"otherNumber":{"type":"number"}},"required":["optionIds"]}},"required":["questionId"]},"minItems":1}},"required":["version","answers"]},"CustomReportSubmitResponseResponse":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"CustomReportViewResponse":{"type":"object","properties":{"id":{"type":"string"},"reportId":{"type":"string"},"version":{"type":"integer"},"respondentId":{"type":"string"},"respondentName":{"type":"string","nullable":true},"submittedAt":{"type":"integer"},"items":{"type":"array","items":{"oneOf":[{"type":"object","properties":{"questionId":{"type":"string"},"query":{"type":"string"},"required":{"type":"boolean"},"kind":{"type":"string","enum":["textField"]},"answer":{"type":"string"}},"required":["questionId","query","kind","answer"]},{"type":"object","properties":{"questionId":{"type":"string"},"query":{"type":"string"},"required":{"type":"boolean"},"kind":{"type":"string","enum":["numberField"]},"answer":{"type":"number"}},"required":["questionId","query","kind","answer"]},{"type":"object","properties":{"questionId":{"type":"string"},"query":{"type":"string"},"required":{"type":"boolean"},"kind":{"type":"string","enum":["choice"]},"answer":{"type":"object","properties":{"options":{"type":"array","items":{"type":"object","properties":{"optionId":{"type":"string"},"label":{"type":"string"},"value":{"type":"string"}},"required":["optionId","label","value"]}},"otherText":{"type":"string"},"otherNumber":{"type":"number"}},"required":["options"]}},"required":["questionId","query","kind","answer"]},{"type":"object","properties":{"questionId":{"type":"string"},"query":{"type":"string"},"required":{"type":"boolean"},"kind":{"type":"string","enum":["choiceMultiSelect"]},"answer":{"type":"object","properties":{"options":{"type":"array","items":{"type":"object","properties":{"optionId":{"type":"string"},"label":{"type":"string"},"value":{"type":"string"}},"required":["optionId","label","value"]}},"otherText":{"type":"string"},"otherNumber":{"type":"number"}},"required":["options"]}},"required":["questionId","query","kind","answer"]}]}}},"required":["id","reportId","version","respondentId","respondentName","items"]}},"parameters":{}},"paths":{"/custom-reports":{"get":{"summary":"List reports (latest version only)","description":"Returns reports available to the requester. Admins receive all published reports. Non-admins receive only reports where both prvilege and SecondPrivileage match the requester. Each item includes only the latest version's questions.","tags":["Custom Reports"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of reports","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomReportListResponse"}}}}}},"post":{"summary":"Create a new custom report (draft)","tags":["Custom Reports"],"security":[{"bearerAuth":[]}],"requestBody":{"description":"Draft report metadata. Questions may be supplied but are optional until publishing.","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomReportCreateRequest"}}}},"responses":{"201":{"description":"Report created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomReportCreateResponse"}}}}}}},"/custom-reports/{id}/publish":{"post":{"summary":"Publish a new version of an existing custom report","tags":["Custom Reports"],"security":[{"bearerAuth":[]}],"parameters":[{"schema":{"type":"string","description":"Custom report identifier"},"required":true,"description":"Custom report identifier","name":"id","in":"path"}],"requestBody":{"description":"Questions that compose the new version. Each publish creates a locked version.","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomReportPublishRequest"}}}},"responses":{"201":{"description":"Version published","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomReportPublishResponse"}}}}}}},"/custom-reports/{id}/responses":{"post":{"summary":"Submit a response for a custom report version","tags":["Custom Reports"],"security":[{"bearerAuth":[]}],"parameters":[{"schema":{"type":"string","description":"Custom report identifier"},"required":true,"description":"Custom report identifier","name":"id","in":"path"}],"requestBody":{"description":"Answers for a specific custom report version.","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomReportSubmitResponseRequest"}}}},"responses":{"201":{"description":"Response submitted","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomReportSubmitResponseResponse"}}}}}}},"/custom-reports/{id}/responses/{responseId}":{"get":{"summary":"Retrieve a submitted response with resolved answers","tags":["Custom Reports"],"security":[{"bearerAuth":[]}],"parameters":[{"schema":{"type":"string","description":"Custom report identifier"},"required":true,"description":"Custom report identifier","name":"id","in":"path"},{"schema":{"type":"string","description":"Response identifier"},"required":true,"description":"Response identifier","name":"responseId","in":"path"}],"responses":{"200":{"description":"Response details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomReportViewResponse"}}}}}}}}}